Portal, Models, Boost and Effects
Created the new boost for both the flip and drift, also made requirements for when the player will get the drift boost.
Made a cookie model, chocolate ball model and a baking cup model.
Made some tire mark effects
Finished the portal to work for both the player and the ai with rotation and position working properly